requirements.txt
alembic==1.16.4
asttokens==3.0.0
choreographer==1.0.9
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
et-xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
greenlet==3.2.3
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.0
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
logistro==1.1.0
mako==1.3.10
markupsafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.48.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openpyxl==3.1.5
optuna==4.4.0
orjson==3.11.1
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.7.1
simplejson==3.20.1
six==1.17.0
sqlalchemy==2.0.41
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
typing-extensions==4.13.2
tzdata==2025.2
wcwidth==0.2.13
